prt_exp_mem(): Move it and related functions to print-dict.c
