+++ link-grammar/prepare/build-disjuncts.c
@@ -344,141 +344,4 @@ GNUC_UNUSED static void print_clause_list(Clause * c)
 		printf("\n");
 	}
 }
-
-/* There is a much better lg_exp_stringify() elsewhere
- * This one is for low-level debug. */
-GNUC_UNUSED void prt_exp(Exp *e, int i)
-{
-	if (e == NULL) return;
-
-	for(int j =0; j<i; j++) printf(" ");
-	printf ("type=%d dir=%c multi=%d cost=%s\n",
-	        e->type, e->dir, e->multi, cost_stringify(e->cost));
-	if (e->type != CONNECTOR_type)
-	{
-		for (e = e->operand_next; e != NULL; e = e->operand_next) prt_exp(e, i+2);
-	}
-	else
-	{
-		for(int j =0; j<i; j++) printf(" ");
-		printf("con=%s\n", e->condesc->string);
-	}
-}
-
-static const char *stringify_Exp_type(Exp_type type)
-{
-	static TLS char unknown_type[32] = "";
-	const char *type_str;
-
-	if (type > 0 && type <= 3)
-	{
-		type_str = ((const char *[]) {"OR", "AND", "CONNECTOR"}) [type-1];
-	}
-	else
-	{
-		snprintf(unknown_type, sizeof(unknown_type), "unknown_type-%d",
-		         (int)(type));
-		type_str = unknown_type;
-	}
-
-	return type_str;
-}
-
-static const char *stringify_Exp_tag(Exp *e, Dictionary dict)
-{
-	static TLS char tag_info[64];
-
-		switch (e->tag_type)
-		{
-			case Exptag_none:
-				return "";
-			case Exptag_dialect:
-				if (dict == NULL)
-				{
-					snprintf(tag_info, sizeof(tag_info), " dialect_tag=%u",
-					         e->tag_id);
-				}
-				else
-				{
-					snprintf(tag_info, sizeof(tag_info), " dialect_tag=%s",
-					       dict->dialect_tag.name[e->tag_id]);
-				}
-				break;
-			case Exptag_macro:
-				if (dict == NULL)
-				{
-					snprintf(tag_info, sizeof(tag_info), " macro_tag");
-				}
-				else
-				{
-					snprintf(tag_info, sizeof(tag_info), " macro_tag=%s",
-					       dict->macro_tag->name[e->tag_id]);
-				}
-				break;
-			default:
-				snprintf(tag_info, sizeof(tag_info), " unknown_tag_type-%d",
-				         (int)(e->tag_type));
-				;
-		}
-
-	return tag_info;
-}
-
-static bool is_ASAN_uninitialized(uintptr_t a)
-{
-	static const uintptr_t asan_uninitialized = (uintptr_t)0xbebebebebebebebeULL;
-
-	return (a == asan_uninitialized);
-}
-
-GNUC_UNUSED void prt_exp_all(Exp *e, int i, Dictionary dict)
-{
-	if (is_ASAN_uninitialized((uintptr_t)e))
-	{
-		printf ("e=UNINITIALIZED\n");
-		return;
-	}
-	if (e == NULL) return;
-
-	for(int j =0; j<i; j++) printf(" ");
-	printf ("e=%p: %s", e, stringify_Exp_type(e->type));
-
-	if (is_ASAN_uninitialized((uintptr_t)e->operand_first))
-		printf(" (UNINITIALIZED operand_first)");
-	if (is_ASAN_uninitialized((uintptr_t)e->operand_next))
-		printf(" (UNINITIALIZED operand_next)");
-
-	if (e->type != CONNECTOR_type)
-	{
-		int operand_count = 0;
-		for (Exp *opd = e->operand_first; NULL != opd; opd = opd->operand_next)
-		{
-			operand_count++;
-			if (is_ASAN_uninitialized((uintptr_t)opd->operand_next))
-			{
-				printf(" (operand %d: UNINITIALIZED operand_next)\n", operand_count);
-				return;
-			}
-		}
-		printf(" (%d operand%s) cost=%s%s\n", operand_count,
-		       operand_count == 1 ? "" : "s", cost_stringify(e->cost),
-		       stringify_Exp_tag(e, dict));
-		for (Exp *opd = e->operand_first; NULL != opd; opd = opd->operand_next)
-		{
-			prt_exp_all(opd, i+2, dict);
-		}
-	}
-	else
-	{
-		printf(" %s%s%c cost=%s%s\n",
-		       e->multi ? "@" : "",
-		       e->condesc ? e->condesc->string : "(condesc=(null))",
-		       e->dir, cost_stringify(e->cost), stringify_Exp_tag(e, dict));
-	}
-}
-
-GNUC_UNUSED static void prt_exp_mem(Exp *e)
-{
-	prt_exp_all(e, 0, NULL);
-}
 #endif /* DEBUG */
